improved decode performance
