+++ srtmlib/core.py
@@ -1,25 +1,13 @@
 #core.py
 
-#mosaic
+# mosaic
 import glob, os
 from collections import defaultdict
 import numpy as np
-import psutil
-from time import time
 import multiprocessing as mp
 
-
-
-#tile
-from struct import *
-
+# tile
 import array
-#import os
-
-#import numpy as np
-#import time
-
-
 
 class Mosaic:
 
@@ -43,8 +31,6 @@ class Mosaic:
 
         self.mosaic_files = []
         self.mosaic_found = {}
-
-        self.pid = os.getpid()
 
     def split_bil(self,reference):
         hemisphere = reference[0]
@@ -138,28 +124,10 @@ class Mosaic:
         print('-------')
         return None
 
-    #get process pid
-    def get_pid(self):
-
-        return self.pid
-
-    #get current memory usage
-    def mem(self, msg=False):
-
-        process = psutil.Process(self.pid)
-        mem = process.memory_info()[0]/float(2**20)
-        if msg:
-            print('Mem usage: {0} MB'.format(mem))
-
-        return mem
-
     def load_tiles(self):
 
-
-        self.mem()
         tiles = []
 
-        start = time()
         for index,item in enumerate(self.mosaic_found.items()):
             if item[1] == True:
                 tiles.append(Tile(self.src_dir+item[0],int(self.arc[0]),index))
@@ -167,19 +135,18 @@ class Mosaic:
         for tile in tiles:
             tile.load_data()
 
-        print('{0:.4f}'.format(time()-start))
+
 
         return None
 
     def load_tiles_mp(self):
-        self.mem()
 
         N = len(self.mosaic_found.items())
 
         procs  = []
         tiles = []
         elements = 0
-        start = time()
+
         # create processes
         for index,item in enumerate(self.mosaic_found.items()):
             if item[1] == True:
@@ -194,12 +161,7 @@ class Mosaic:
         for p in procs:
             p.join()
 
-        print('{0:.4f}s'.format(time()-start))
-
         return None
-
-
-
 
 
 class Tile:
@@ -220,23 +182,9 @@ class Tile:
             self.shape = (0,0)
 
 
-    def read_bytes_LC(self, bytes_to_read):
-        # DEPRECATED
-        # ---------
-        # read binary file as unsigned char type
-        # using a list comprehension
-
-        f = open(self.src_file, "rb")
-        raw_bytes = [ unpack('B',f.read(1))[0] for byte in range(bytes_to_read) ]
-        f.close()
-        loaded_bytes =  np.asarray(raw_bytes)
-
-        return loaded_bytes
-
-
     def read_bytes(self, bytes_to_read):
         # read binary file as unsigned char type
-        # using an array.fromfile
+        # return 1D np.array
 
         raw_bytes = array.array('B')
 
@@ -247,33 +195,31 @@ class Tile:
 
         return loaded_bytes
 
-
     def decode_hgt(self, src_bytes, wrapped):
 
-        # decode height stored in big endian format
-        size = src_bytes.shape[0]
-        heights = []
+        # split array into high and low bytes
+        low_b  = src_bytes[::2]
+        high_b = src_bytes[1::2]
+        heights = None
 
         if not wrapped:
-            heights = [ ((src_bytes[i+1]<<8) + src_bytes[i]) for i in range(0, size, 2) ]
+            heights = np.left_shift(high_b,8) + low_b
         else:
-            heights = [ src_bytes[i] for i in range(0, size, 2) ]
-        decoded = np.asarray(heights)
+            heights = low_b
 
-        return decoded
+        return heights
 
 
     def load_data(self, wrapped=False):
 
         samples_per_line, lines = self.shape
         bytes_per_sample = 2
-        max_val = 8480
 
-        # read from file
-        # decode data
         heights = self.decode_hgt(self.read_bytes(samples_per_line*bytes_per_sample*lines), wrapped)
+
         # special height values handling
-        # TODO
+        # TO DO
+        # max_val = 8480
         # self.heights = np.clip( np.reshape(heights,(samples_per_line,lines)),0,max_val)
 
         self.heights = np.reshape(heights,(samples_per_line,lines))
